Перенесен вызов экрана Onboarding в настройки
