+++ lib/ui/sight/sight_list/sight_list_screen.dart
@@ -8,7 +8,6 @@ import 'package:places/ui/components/icon_action.dart';
 import 'package:places/ui/components/main_gradient_overlay.dart';
 import 'package:places/ui/icons/menu/svg_icons.dart';
 import 'package:places/ui/icons/svg_icons.dart';
-import 'package:places/ui/onboarding/onboarding_screen.dart';
 import 'package:places/ui/sight/edit_sight/add_sight_screen.dart';
 import 'package:places/ui/sight/filters/filters_state.dart';
 import 'package:places/ui/sight/search/sight_search_screen.dart';
@@ -54,26 +53,9 @@ class _SightListScreenState extends State<SightListScreen> {
         highlightElevation: 0,
         icon: const _FloatingButtonText(),
         label: const SizedBox(),
-        onPressed: openOnboardingScreen,
+        onPressed: openAddSightScreen,
         // onPressed: openAddSightScreen,
       ),
-    );
-  }
-
-  void openOnboardingScreen() {
-    showModalBottomSheet<void>(
-      context: context,
-      isScrollControlled: true,
-      builder: (context) {
-        return OnboardingScreen(
-          onSkip: () {
-            Navigator.of(context).pop();
-          },
-          onStart: () {
-            Navigator.of(context).pop();
-          },
-        );
-      },
     );
   }
